+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Use `currency` instead of not existing property `currencyCode` from runtime.
+
 ## [0.1.0] - 2019-07-15
 
 ### Removed